Tidy up test fixtures
